Fixed patient summary and timeline styles
