+++ app/styles/_patients.scss
@@ -162,7 +162,7 @@
             position: absolute;
             top: 0;
             left: -35px;
-            font-size: 42px;
+            font-size: 38px;
             line-height: 55px;
             border: 5px solid $brand-secondary;
             border-radius: 50%;
